// solana/ts/sdk/anchor-idl/1_0_0/dummy_transfer_hook.d.ts
export type DummyTransferHook = {
    "version": "0.1.0";
    "name": "dummy_transfer_hook";
    "instructions": [
        {
            "name": "initializeExtraAccountMetaList";
            "accounts": [
                {
                    "name": "payer";
                    "isMut": true;
                    "isSigner": true;
                },
                {
                    "name": "extraAccountMetaList";
                    "isMut": true;
                    "isSigner": false;
                },
                {
                    "name": "mint";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "tokenProgram";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "associatedTokenProgram";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "counter";
                    "isMut": true;
                    "isSigner": false;
                },
                {
                    "name": "systemProgram";
                    "isMut": false;
                    "isSigner": false;
                }
            ];
            "args": [];
        },
        {
            "name": "transferHook";
            "accounts": [
                {
                    "name": "sourceToken";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "mint";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "destinationToken";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "authority";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "extraAccountMetaList";
                    "isMut": false;
                    "isSigner": false;
                },
                {
                    "name": "dummyAccount";
                    "isMut": false;
                    "isSigner": false;
                    "docs": [
                        "computes and the on-chain code correctly passes on the PDA."
                    ];
                },
                {
                    "name": "counter";
                    "isMut": true;
                    "isSigner": false;
                }
            ];
            "args": [
                {
                    "name": "amount";
                    "type": "u64";
                }
            ];
        }
    ];
    "accounts": [
        {
            "name": "counter";
            "type": {
                "kind": "struct";
                "fields": [
                    {
                        "name": "count";
                        "type": "u64";
                    }
                ];
            };
        }
    ];
};
export declare const IDL: DummyTransferHook;
//# sourceMappingURL=dummy_transfer_hook.d.ts.map